+++ src/sdk.ts
@@ -1,20 +1,16 @@
+import type { Domain, SubDomain } from "./types.ts";
 import {
-  type Domain,
   type DomainStorageValue,
-  KeyValueDbContract,
-  type Record,
-  RecordKey,
-  RecordType,
-  RegistryContract,
-  ReverseRegistrarContract,
+  KEY_VALUE_DB_CONTRACT,
+  REGISTRY_CONTRACT,
+  REVERSE_REGISTRAR_CONTRACT,
   SIMULATION_ACCOUNT,
   type SorobanDomainsSDKParams,
-  type SubDomain,
 } from "./types.ts";
+import { Buffer } from "buffer";
 import { crypto } from "@std/crypto";
-import { encodeHex } from "@std/encoding";
 import { concat } from "@std/bytes";
-import { Domain404Error, DomainData404Error, DomainDataUnsupportedValueType, ReverseDomain404Error } from "./errors.ts";
+import { DomainData404Error, DomainDataUnsupportedValueType, ReverseDomain404Error } from "./errors.ts";
 import {
   Account,
   Contract,
@@ -27,16 +23,20 @@ import {
   xdr,
 } from "@stellar/stellar-sdk";
 import { decodeHex } from "@std/encoding/hex";
+import { RegistryV2Client } from "./apis/mod.ts";
 
 export class SorobanDomainsSDK {
   readonly #rpcUrl?: string;
+  readonly #allowHttp?: boolean;
   readonly #simulationAccount: string;
   readonly #globalFee: string;
-  readonly #registerContract: string;
+  readonly #registryContract: string;
   readonly #keyValueDbContract: string;
   readonly #reverseRegistrarContract: string;
   readonly #network: Networks;
   readonly #defaultTimeout: number;
+
+  readonly registryV2Client: RegistryV2Client;
 
   get server(): rpc.Server {
     if (!this.#rpcUrl) {
@@ -46,42 +46,54 @@ export class SorobanDomainsSDK {
     return new rpc.Server(this.#rpcUrl);
   }
 
-  constructor(params: SorobanDomainsSDKParams) {
-    this.#rpcUrl = params.rpcUrl;
+  constructor(params: SorobanDomainsSDKParams = {}) {
+    this.#rpcUrl = params.rpcUrl || "https://rpc.lightsail.network";
+    this.#allowHttp = !!params.allowHttp;
     this.#simulationAccount = params.simulationAccount || SIMULATION_ACCOUNT;
-    this.#globalFee = params.defaultFee || "100";
-    this.#registerContract = params.registryContractId || RegistryContract.v0;
-    this.#keyValueDbContract = params.keyValuesDatabaseContractId || KeyValueDbContract.v0;
-    this.#reverseRegistrarContract = params.reverseRegistrarContractId || ReverseRegistrarContract.v0;
+    this.#globalFee = params.defaultFee || "100000";
+    this.#registryContract = params.registryContractId || REGISTRY_CONTRACT;
+    this.#keyValueDbContract = params.keyValuesDatabaseContractId || KEY_VALUE_DB_CONTRACT;
+    this.#reverseRegistrarContract = params.reverseRegistrarContractId || REVERSE_REGISTRAR_CONTRACT;
     this.#network = params.network || Networks.PUBLIC;
     this.#defaultTimeout = params.defaultTimeout || 0;
+
+    this.registryV2Client = new RegistryV2Client({
+      contractId: this.#registryContract,
+      networkPassphrase: this.#network,
+      rpcUrl: this.#rpcUrl,
+      allowHttp: this.#allowHttp,
+    });
   }
 
   static hash(text: string | Uint8Array): Uint8Array {
-    return new Uint8Array(
-      crypto.subtle.digestSync("KECCAK-256", typeof text === "string" ? new TextEncoder().encode(text) : text),
-    );
+    const data = (typeof text === "string" ? new TextEncoder().encode(text) : text) as Uint8Array<ArrayBuffer>;
+    return new Uint8Array(crypto.subtle.digestSync("KECCAK-256", data));
   }
 
   /**
    * This function takes a domain and generates the "node" value of the parsed domain.
-   * This "node" value can be used to fetch data from the contract (either if is a Record or a SubRecord)
+   * This "node" value can be used to fetch data from the contract
+   *
+   * You need to provide the chunks in reverse, so for example the domain "treasury.addresses.protocol.xlm" the params will be:
+   * - domainChunks: the bytes in the order ["protocol", "addresses","treasury"]
+   * - tld: "xlm"
    */
-  static parseDomain(params: { domain: string; subDomain?: string; tld?: string }): string {
-    const node: Uint8Array = this.hash(concat([
-      this.hash(params.tld || "xlm"),
-      this.hash(params.domain),
-    ]));
-
-    if (params.subDomain) {
-      const subNode: Uint8Array = this.hash(concat([
-        this.hash(node),
-        this.hash(params.subDomain),
-      ]));
-      return encodeHex(subNode);
-    } else {
-      return encodeHex(node);
+  static generateNode(domainChunks: Array<Uint8Array | string>, tld: Uint8Array | string): Uint8Array {
+    if (domainChunks.length < 1) {
+      throw new Error("Domain chunks must be at least size 1");
     }
+    const chunks = [...domainChunks];
+    let node: Uint8Array = this.hash(concat([
+      this.hash(tld),
+      this.hash(chunks.shift()!),
+    ]));
+    while (chunks.length > 0) {
+      node = this.hash(concat([
+        this.hash(node),
+        this.hash(chunks.shift()!),
+      ]));
+    }
+    return node;
   }
 
   /**
@@ -94,71 +106,282 @@ export class SorobanDomainsSDK {
     const domainRegex: RegExp = new RegExp("^[a-z]+(\\.[a-z]+)*\\.[a-z]{2,}$");
     if (!domainRegex.test(domain)) return false;
     const parts: string[] = domain.split(".");
-    if (parts.length > 3) return false;
-    return parts.every((part: string): boolean => part.length <= 15);
+    return parts.every((part: string): boolean => part.length <= 15 && part.length >= 1);
   }
 
   /**
-   * This function parses the domain you want to use, and it will search it in the contract.
-   * This function doesn't validate the domain you're providing is a valid one.
+   * This method will take a domain, validate it and generate the reversed chunks this library use.
+   *
+   * @param domain {String} - The domain to separate in chunks
    */
-  async searchDomain(params: { domain: string; subDomain?: string }): Promise<Record> {
-    const domainNode: string = SorobanDomainsSDK.parseDomain({
-      domain: params.domain.toLocaleLowerCase(),
-      subDomain: params.subDomain?.toLocaleLowerCase(),
-    });
+  static generateDomainChunks(domain: string): string[] {
+    this.isValidDomain(domain);
+    const domainChunks: string[] = domain.split(".").toReversed();
+    if (domainChunks.length < 2) {
+      throw new Error(
+        "Domain is not long enough, remember that you need to include at least the TLD and the root domain",
+      );
+    }
+    return domainChunks;
+  }
 
-    const contract: Contract = new Contract(this.#registerContract);
-    const nodeBytes: xdr.ScVal = xdr.ScVal.scvBytes(decodeHex(domainNode));
+  /**
+   * This method is used to register a new domain or subdomain.
+   *
+   * @param params {Object}
+   * @param params.domain {String} - The domain to register, for example: hello.mate.xlm
+   * @param params.owner {String|Undefined} - The account that will receive the NFD token; Not needed if you are registering a subdomain, if not provided then the "address" will be used
+   * @param params.address {String} - The address to which this domain will translate to
+   * @param params.periods {String} - Amount of years the domain will be registered
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async register(params: {
+    domain: string;
+    owner?: string;
+    address: string;
+    periods?: bigint;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(params.domain);
+    const tld: string = domainChunks.shift()!;
+    let tx: Transaction;
+    if (domainChunks.length > 1) {
+      const subdomain: string = domainChunks.pop()!;
+      const parentNode: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+      const { raw } = await this.registryV2Client.register_sub({
+        address: params.address,
+        new_subdomain: Buffer.from(subdomain),
+        parent: {
+          values: [Buffer.from(parentNode)],
+          tag: domainChunks.length > 1 ? "SubDomain" : "Domain",
+        },
+      }, {
+        fee: this.#globalFee,
+        simulate: false,
+        timeoutInSeconds: 0,
+      });
+      tx = raw!.build();
+    } else {
+      const { raw } = await this.registryV2Client.register({
+        new_domain: Buffer.from(domainChunks[0]),
+        tld: Buffer.from(tld),
+        address: params.address,
+        owner: params.owner || params.address,
+        periods: params.periods || 1n,
+      }, {
+        fee: this.#globalFee,
+        simulate: false,
+        timeoutInSeconds: 0,
+      });
+      tx = raw!.build();
+    }
+    return this.server.prepareTransaction(tx);
+  }
 
-    const record_key: xdr.ScVal = xdr.ScVal.scvVec([
-      xdr.ScVal.scvSymbol(params.subDomain ? RecordKey.SubRecord : RecordKey.Record),
-      nodeBytes,
-    ]);
-
-    const transaction: Transaction = new TransactionBuilder(new Account(this.#simulationAccount, "0"), {
-      networkPassphrase: this.#network,
+  /**
+   * This method is used to update the address at which a domain points to.
+   *
+   * @param params {Object}
+   * @param params.domain {String} - The domain we will update, for example: hello.mate.xlm
+   * @param params.newAddress {String} - The new address
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async updateAddress(params: {
+    domain: string;
+    newAddress: string;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(params.domain);
+    const tld: string = domainChunks.shift()!;
+    const node: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+    const { raw } = await this.registryV2Client.update_address({
+      new_address: params.newAddress,
+      record_key: {
+        values: [Buffer.from(node)],
+        tag: domainChunks.length > 1 ? "SubDomain" : "Domain",
+      },
+    }, {
       fee: this.#globalFee,
-    })
-      .setTimeout(this.#defaultTimeout)
-      .addOperation(contract.call("record", record_key))
+      simulate: false,
+      timeoutInSeconds: 0,
+    });
+    return this.server.prepareTransaction(raw!.build());
+  }
+
+  /**
+   * This method is used to extend the registration period for a domain. This can be called by anyone and extend a domain.
+   * Be aware that if you extend a domain that it is not near the expiration date, you will only extend it from the moment you call the method, not including any possible period left that was already paid.
+   *
+   * @param params {Object}
+   * @param params.payer {String} - The account that will pay for the extension of the domain registration
+   * @param params.domain {String} - The domain to extend, it needs tobe only the root domain and the tld. For example sorobandomains.xlm
+   * @param params.periods {String} - Amount of years the domain will be extended
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async renew(params: {
+    payer: string;
+    domain: string;
+    periods: number;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(params.domain);
+    const tld: string = domainChunks.shift()!;
+    const node: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+    const { raw } = await this.registryV2Client.renew({
+      caller: params.payer,
+      node: Buffer.from(node),
+      periods: BigInt(params.periods),
+    }, {
+      fee: this.#globalFee,
+      simulate: false,
+      timeoutInSeconds: 0,
+    });
+    return this.server.prepareTransaction(raw!.build());
+  }
+
+  /**
+   * This method can be called by someone to claim a domain that had been 30 days expired.
+   *
+   * @param params {Object}
+   * @param params.buyer {String} - The account that will pay to claim the domain
+   * @param params.domain {String} - The domain to extend, it needs tobe only the root domain and the tld. For example sorobandomains.xlm
+   * @param params.address {String} - The address to which this domain will translate to
+   * @param params.periods {String} - Amount of years the domain will be extended
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async claim(params: {
+    buyer: string;
+    domain: string;
+    address: string;
+    periods: number;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(params.domain);
+    const tld: string = domainChunks.shift()!;
+    const node: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+    const { raw } = await this.registryV2Client.claim({
+      caller: params.buyer,
+      node: Buffer.from(node),
+      address: params.address,
+      periods: BigInt(params.periods),
+    }, {
+      fee: this.#globalFee,
+      simulate: false,
+      timeoutInSeconds: 0,
+    });
+    return this.server.prepareTransaction(raw!.build());
+  }
+
+  /**
+   * This method is used by the admin to burn a domain.
+   *
+   * @param params {Object}
+   * @param params.domain {String} - The domain to extend, it needs tobe only the root domain and the tld. For example sorobandomains.xlm
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async evict(params: {
+    domain: string;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(params.domain);
+    const tld: string = domainChunks.shift()!;
+    const node: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+    const { raw } = await this.registryV2Client.evict({
+      node: Buffer.from(node),
+    }, {
+      fee: this.#globalFee,
+      simulate: false,
+      timeoutInSeconds: 0,
+    });
+    return this.server.prepareTransaction(raw!.build());
+  }
+
+  /**
+   * This method is used to migrate a domain from the Registry V1 to this new Registry.
+   *
+   * @param params {Object}
+   * @param params.domain {String} - The root domain to migrate, for example: mate.xlm
+   * @param params.source {String} - The Stellar account which will sign and pay for the Transaction fee
+   */
+  async migrateFromV1(params: {
+    domain: string;
+    tld: string;
+    source: string;
+  }): Promise<Transaction> {
+    this.registryV2Client.options.publicKey = params.source;
+    const [tld, domain]: string[] = params.domain.split(".").toReversed().map((item: string): string =>
+      item.trim().toLowerCase()
+    );
+    const { raw } = await this.registryV2Client.migrate({
+      new_domain: Buffer.from(domain),
+      tld: Buffer.from(tld),
+    }, {
+      fee: this.#globalFee,
+      simulate: false,
+      timeoutInSeconds: 0,
+    });
+    return this.server.prepareTransaction(raw!.build());
+  }
+
+  /**
+   * @param domain {String} - The domain we will get from the registry, for example: my.mate.xlm
+   */
+  async searchDomain<Domain>(domain: string): Promise<Domain>;
+  async searchDomain<SubDomain>(domain: string): Promise<SubDomain>;
+  async searchDomain(domain: string): Promise<Domain | SubDomain> {
+    const domainChunks: string[] = SorobanDomainsSDK.generateDomainChunks(domain);
+    const tld: string = domainChunks.shift()!;
+    const node: Uint8Array = SorobanDomainsSDK.generateNode(domainChunks, tld);
+    const tx: Transaction = new TransactionBuilder(new Account(SIMULATION_ACCOUNT, "0"), {
+      networkPassphrase: this.#network,
+      fee: "0",
+    }).setTimeout(0)
+      .addOperation(
+        new Contract(this.#registryContract).call(
+          "record",
+          xdr.ScVal.scvVec([
+            xdr.ScVal.scvSymbol(domainChunks.length > 1 ? "SubDomain" : "Domain"),
+            xdr.ScVal.scvBytes(node),
+          ]),
+        ),
+      )
       .build();
 
-    const sim: rpc.Api.SimulateTransactionResponse = await this.server.simulateTransaction(transaction);
+    const sim = await this.server.simulateTransaction(tx);
 
     if (rpc.Api.isSimulationError(sim)) {
       throw new Error(sim.error);
     }
 
-    const result = scValToNative(sim.result!.retval);
-
-    if (!result) {
-      throw new Domain404Error();
+    if (rpc.Api.isSimulationSuccess(sim)) {
+      const [domain, subDomain] = scValToNative(sim.result!.retval);
+      if (subDomain) {
+        return {
+          address: subDomain.address,
+          domain: subDomain.domain.toString(),
+          parent: subDomain.parent.toString("hex"),
+          root: subDomain.root.toString("hex"),
+          node: subDomain.node.toString("hex"),
+          snapshot: subDomain.snapshot,
+        };
+      } else {
+        return {
+          address: domain.address,
+          domain: domain.domain.toString(),
+          tld: domain.tld.toString(),
+          node: domain.node.toString("hex"),
+          snapshot: domain.snapshot,
+          token_id: domain.token_id,
+        };
+      }
     }
 
-    if (result[0] === "Domain") {
-      return {
-        type: RecordType.Domain,
-        value: {
-          node: result[1].node.toString("hex"),
-          owner: result[1].owner.toString("hex"),
-          address: result[1].address,
-          exp_date: result[1].exp_date.toString(),
-          snapshot: result[1].snapshot.toString(),
-          collateral: result[1].collateral.toString(),
-        } satisfies Domain,
-      };
-    } else {
-      return {
-        type: RecordType.SubDomain,
-        value: {
-          node: result[1].node.toString("hex"),
-          parent: result[1].parent.toString("hex"),
-          address: result[1].address,
-          snapshot: result[1].snapshot.toString(),
-        } satisfies SubDomain,
-      };
-    }
+    throw new Error("Unexpected Error, please contact support");
   }
 
   async getDomainData(params: { node: string; key: string }): Promise<DomainStorageValue> {